feat: Adiciona novos atributos de geolocalização e região nas entidades City e State
