--- src/main/java/com/querodoar/querodoar_api/city/City.java
@@ -1,64 +1,0 @@
-package com.querodoar.querodoar_api.city;
-
-import com.querodoar.querodoar_api.city.entity.State;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-
-import java.math.BigDecimal;
-
-@Entity
-@Table(name = "city")
-public class City {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "city_id", nullable = false)
-    private Integer id;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "state_id", nullable = false)
-    private State state;
-
-    @Size(max = 60)
-    @NotNull
-    @Column(name = "name", nullable = false, length = 60)
-    private String name;
-
-    @NotNull
-    @Column(name = "ibge_code", nullable = false, precision = 5)
-    private BigDecimal ibgeCode;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public State getState() {
-        return state;
-    }
-
-    public void setState(State state) {
-        this.state = state;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public BigDecimal getIbgeCode() {
-        return ibgeCode;
-    }
-
-    public void setIbgeCode(BigDecimal ibgeCode) {
-        this.ibgeCode = ibgeCode;
-    }
-
-}
